backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-multipart>=0.0.18
Pillow>=11.0.0
openai>=1.60.0
google-genai>=1.0.0
httpx>=0.28.0
trafilatura>=2.0.0
beautifulsoup4>=4.12.0
google-search-results>=2.4.2
streamlit>=1.41.0
